ml-model/pyproject.toml
[tool.poetry]
name = "ml-model"
version = "0.1.0"
description = "Environment to train and serve models"
authors = ["Luis Chahua <luischahua929@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pandas = {extras = ["fss", "gcp"], version = "^2.0.3"}
lightgbm = "^3.3.5"
scikit-learn = "^1.3.0"
joblib = "^1.2"
utilities = { path = "./utilities", develop = true }

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = ">=7.3.1,<8.0.0"

[tool.poetry.group.notebook]
optional = true

[tool.poetry.group.notebook.dependencies]
ipykernel = "^6.24.0"
matplotlib = "^3.7.1"
Jinja2 = ">=3.0.1,<4.0.0"
google-cloud-bigquery = "^2.30.0"
graphviz = "^0.20.1"

[tool.poetry.group.serving]
optional = true

[tool.poetry.group.serving.dependencies]
fastapi = {extras = ["uvicorn"], version = "^0.100.0"}
uvicorn = "^0.23.1"
google-cloud-storage = "^2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"